Update JYCardTrick.java
My previous button wouldn't show up on the frame on the updated version of the game, however i came across a new approach, once the game is finished, to close after click on okay, it would take the user back to the main menu. Is there a way to access the label of the Okay button and rename it? or perhaps add another button or so?